mapper/requirements.txt
google-cloud-storage
requests
numpy